blockchain: Reorganize exports
